# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API para el sistema interno de Mariana Textil.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clientes/resumen:
    get:
      operationId: getClientesResumen
      tags: [clientes]
      summary: Resumen de cartera (clientes_finanzas requerido)
      responses:
        "200":
          description: Resumen de cartera
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientesResumen"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: Lista el catálogo operativo de clientes (sin datos financieros)
      responses:
        "200":
          description: Lista de clientes operativos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Da de alta un cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteInput"
      responses:
        "201":
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /clientes/{id}:
    get:
      operationId: getCliente
      tags: [clientes]
      summary: Obtiene detalle operativo de un cliente
      parameters:
        - $ref: "#/components/parameters/ClienteId"
      responses:
        "200":
          description: Detalle del cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateCliente
      tags: [clientes]
      summary: Actualiza datos operativos de un cliente
      parameters:
        - $ref: "#/components/parameters/ClienteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteUpdate"
      responses:
        "200":
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /clientes/{id}/credito:
    get:
      operationId: getClienteCredito
      tags: [clientes]
      summary: Semáforo de crédito del cliente (clientes_credito requerido)
      parameters:
        - $ref: "#/components/parameters/ClienteId"
      responses:
        "200":
          description: Datos de crédito del cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteCredito"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /clientes/{id}/precios:
    get:
      operationId: getClientePrecios
      tags: [clientes]
      summary: Precios unitarios negociados por producto (clientes_precios requerido)
      parameters:
        - $ref: "#/components/parameters/ClienteId"
      responses:
        "200":
          description: Historial de precios unitarios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientePrecios"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /clientes/{id}/estado-cuenta:
    get:
      operationId: getClienteEstadoCuenta
      tags: [clientes]
      summary: Estado de cuenta del cliente (clientes_finanzas requerido)
      parameters:
        - $ref: "#/components/parameters/ClienteId"
      responses:
        "200":
          description: Estado de cuenta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteEstadoCuenta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /clientes/{id}/compras:
    get:
      operationId: getClienteCompras
      tags: [clientes]
      summary: Historial de compras del cliente (clientes_finanzas requerido)
      parameters:
        - $ref: "#/components/parameters/ClienteId"
      responses:
        "200":
          description: Compras del cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteCompras"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /clientes/{id}/estadisticas:
    get:
      operationId: getClienteEstadisticas
      tags: [clientes]
      summary: Estadísticas analíticas del cliente (clientes_finanzas requerido)
      parameters:
        - $ref: "#/components/parameters/ClienteId"
      responses:
        "200":
          description: Estadísticas del cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteEstadisticas"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /clientes/{id}/pagos:
    get:
      operationId: getClientePagos
      tags: [clientes]
      summary: Pagos del cliente (clientes_finanzas requerido)
      parameters:
        - $ref: "#/components/parameters/ClienteId"
      responses:
        "200":
          description: Pagos del cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientePagos"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createClientePago
      tags: [clientes]
      summary: Registra un pago de cliente (clientes_finanzas + crear requerido)
      parameters:
        - $ref: "#/components/parameters/ClienteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientePagoInput"
      responses:
        "201":
          description: Pago registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientePago"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    ClienteId:
      name: id
      in: path
      required: true
      schema:
        type: number
  responses:
    Unauthorized:
      description: Sesión no válida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Sin permisos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Cliente:
      type: object
      required: [id, nombre, activo, createdAt, updatedAt]
      properties:
        id: { type: number }
        nombre: { type: string }
        telefono: { type: ["string", "null"] }
        correo: { type: ["string", "null"] }
        direccion: { type: ["string", "null"] }
        rfc: { type: ["string", "null"] }
        notas: { type: ["string", "null"] }
        activo: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ClienteInput:
      type: object
      required: [nombre]
      properties:
        nombre: { type: string, minLength: 1, maxLength: 200 }
        telefono: { type: ["string", "null"] }
        correo: { type: ["string", "null"] }
        direccion: { type: ["string", "null"] }
        rfc: { type: ["string", "null"] }
        notas: { type: ["string", "null"] }
    ClienteUpdate:
      type: object
      properties:
        nombre: { type: string, minLength: 1, maxLength: 200 }
        telefono: { type: ["string", "null"] }
        correo: { type: ["string", "null"] }
        direccion: { type: ["string", "null"] }
        rfc: { type: ["string", "null"] }
        notas: { type: ["string", "null"] }
        activo: { type: boolean }
    ClienteCredito:
      type: object
      required: [clienteId, limiteCredito, saldoActual, creditoDisponible, puedeComprarCredito]
      properties:
        clienteId: { type: number }
        limiteCredito: { type: string }
        saldoActual: { type: string }
        creditoDisponible: { type: string }
        puedeComprarCredito: { type: boolean }
    ClientePrecios:
      type: object
      required: [clienteId, precios]
      properties:
        clienteId: { type: number }
        precios:
          type: array
          items:
            type: object
            required: [productoId, sku, precioUnitario, fecha, promedio3]
            properties:
              productoId: { type: number }
              sku: { type: string }
              precioUnitario: { type: string }
              fecha: { type: string, format: date }
              promedio3: { type: string, description: "Promedio de las últimas 3 compras de ese producto" }
        nota: { type: ["string", "null"] }
    ClienteMovimiento:
      type: object
      properties:
        tipo: { type: string }
        importe: { type: string }
        fecha: { type: string, format: date-time }
        notas: { type: ["string", "null"] }
    ClienteEstadoCuenta:
      type: object
      required: [clienteId, movimientos, saldoActual]
      properties:
        clienteId: { type: number }
        movimientos:
          type: array
          items:
            $ref: "#/components/schemas/ClienteMovimiento"
        saldoActual: { type: string }
    ClienteCompraItem:
      type: object
      properties:
        id: { type: number }
        fecha: { type: string, format: date-time }
        total: { type: string }
    ClienteCompras:
      type: object
      required: [clienteId, compras, total]
      properties:
        clienteId: { type: number }
        compras:
          type: array
          items:
            $ref: "#/components/schemas/ClienteCompraItem"
        total: { type: number }
    ClienteEstadisticas:
      type: object
      required: [clienteId]
      properties:
        clienteId: { type: number }
        totalCompras: { type: ["string", "null"] }
        comprasCount: { type: ["number", "null"] }
    ClientePagoItem:
      type: object
      properties:
        id: { type: number }
        importe: { type: string }
        fecha: { type: string, format: date-time }
        formaPago: { type: ["string", "null"] }
    ClientePagos:
      type: object
      required: [clienteId, pagos]
      properties:
        clienteId: { type: number }
        pagos:
          type: array
          items:
            $ref: "#/components/schemas/ClientePagoItem"
    ClientePago:
      type: object
      required: [id, clienteId]
      properties:
        id: { type: number }
        clienteId: { type: number }
    ClientePagoInput:
      type: object
      required: [importe, formaPago]
      properties:
        importe: { type: number, minimum: 0.01 }
        formaPago: { type: string }
        referencia: { type: ["string", "null"] }
        notas: { type: ["string", "null"] }
    ClientesResumen:
      type: object
      required: [totalClientes, clientesConSaldo, totalCartera, totalVencido]
      properties:
        totalClientes: { type: number }
        clientesConSaldo: { type: number }
        totalCartera: { type: string }
        totalVencido: { type: string }
...