# lib/api-zod/src/generated/types/clienteCompras.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para el sistema interno de Mariana Textil.
 * OpenAPI spec version: 0.1.0
 */
import type { ClienteCompraItem } from './clienteCompraItem';

export interface ClienteCompras {
  clienteId: number;
  compras: ClienteCompraItem[];
  total: number;
}