# lib/api-zod/src/generated/types/clienteCredito.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para el sistema interno de Mariana Textil.
 * OpenAPI spec version: 0.1.0
 */

export interface ClienteCredito {
  clienteId: number;
  limiteCredito: string;
  saldoActual: string;
  creditoDisponible: string;
  puedeComprarCredito: boolean;
}