# lib/api-zod/src/generated/types/clienteEstadisticas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para el sistema interno de Mariana Textil.
 * OpenAPI spec version: 0.1.0
 */

export interface ClienteEstadisticas {
  clienteId: number;
  /** @nullable */
  totalCompras?: string | null;
  /** @nullable */
  comprasCount?: number | null;
}