# lib/api-zod/src/generated/types/clienteEstadoCuenta.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para el sistema interno de Mariana Textil.
 * OpenAPI spec version: 0.1.0
 */
import type { ClienteMovimiento } from './clienteMovimiento';

export interface ClienteEstadoCuenta {
  clienteId: number;
  movimientos: ClienteMovimiento[];
  saldoActual: string;
}