# lib/api-zod/src/generated/types/clientePagoItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para el sistema interno de Mariana Textil.
 * OpenAPI spec version: 0.1.0
 */

export interface ClientePagoItem {
  id?: number;
  importe?: string;
  fecha?: Date;
  /** @nullable */
  formaPago?: string | null;
}